+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ChangeLog
 
+## Unreleased
+
+### Added
+- added table of contents with songs title and arrangement to song sheets
+
+### Changed
+- updated external dependencies (certifi, pydantic)
+
 ## 0.10.10 (2025-10-04)
 
 ### Added